Change datatype of rate from int to double

+++ dubins_trajectory_generator/src/dubins_trajectory_node.cpp
@@ -9,8 +9,8 @@ int main(int argc, char** argv) {
     ros::NodeHandle nh_private("~");
 
     DubinsTrajectory trajectory(nh, nh_private);
-    int rate;
-    nh_private.param("rate",rate,10);
+    double rate;
+    nh_private.param("rate",rate,10.0);
 
     ros::Rate loop_rate(rate);
